requirement.txt
langchain
openai
torch
sentence-transformers
pinecone-client
flask
psycopg2
flask-cors
langchain-openai